only start the servers if they arent already running

+++ rspec/spec/runner.rb
@@ -1,17 +1,21 @@
 require 'httparty'
 
 class Runner
+  API_PORT = 3000
+  FRONTEND_PORT = 4200
+
   @@pid = nil
+  @@frontend_pid = nil
 
   def self.run
-    @@pid = Process.spawn('cd ../server/ && npm start', {pgroup: true})
-    @@frontend_pid = Process.spawn('cd ../client/ && ng serve', {pgroup: true})
+    @@pid = Process.spawn('cd ../server/ && npm start', {pgroup: true}) unless `lsof -Pi :#{API_PORT} -sTCP:LISTEN -t`
+    @@frontend_pid = Process.spawn('cd ../client/ && ng serve', {pgroup: true}) unless `lsof -Pi :#{FRONTEND_PORT} -sTCP:LISTEN -t`
 
     number_of_attempts = 0
     while number_of_attempts < 10
       begin
-        response = HTTParty.get('http://localhost:3000/ping') # hit the ping endpoint
-        frontend_response = HTTParty.get('http://localhost:4200')
+        response = HTTParty.get("http://localhost:#{API_PORT}/ping") # hit the ping endpoint
+        frontend_response = HTTParty.get("http://localhost:#{FRONTEND_PORT}")
 
         break if response.code == 200 && frontend_response.code == 200 # success!
       rescue Errno::ECONNREFUSED # rescue only if the exception is a failure to connect
@@ -25,7 +29,7 @@ class Runner
 
   def self.stop
     # kill the processes
-    Process.kill(-9, @@pid)
-    Process.kill(-9, @@frontend_pid)
+    Process.kill(-9, @@pid) if @@pid
+    Process.kill(-9, @@frontend_pid) if @@frontend_pid
   end
 end
